chore: moved class to correct package
